requirements.txt
gcsfs
yfinance
pandas
numpy
matplotlib
joblib
scikit-learn
pyyaml
kafka-python
streamlit
apache-airflow==2.8.1
xgboost==1.7.6
google-cloud-storage
pathlib
datetime
altair
plotly
tqdm
feedparser